requirements.txt
Pillow==6.1.0
python-magic==0.4.15
tabulate==0.8.3
